Add files via upload
Welcome!

Thanks for checking out this front-end coding challenge.
To do this challenge,you need a basic understanding of HTML and CSS.

The Challenge
--------------------------------------------------------

Your challenge is to build out this landing page and get it looking as close to the design as possible.

You can use any tools you like to help you complete the challenge.So if you've got something you'd like to practice,feel free to give it a go.

Your users should be able to:

* View the optimal layout for the site depending on their device's screen size

* See hover states for all interactive elements on the page


Where to find everything
--------------------------------------------------------

Your task is to build out the project to the designs inside the /design folder. You will find both a mobile and a desktop version of the design to work to.

The designs are in JPG static format. This will mean that you'll need to use your best judgment for styles such as font-size,padding and margin.This should help train your eye to perceive differences in spacings and sizes.